split apart metadata tests
